Extracted tsx logic.
